Replace ROM calls SNDON, SNDOFF, and SETMUX
The ROM calls for SNDON, SNDOFF, and SETMUX are quite trivial so we now
provide our own routines. These routines take up slightly more space
than the ROM routines because the tricks to make the code smaller are
not actually required. That also means that SETMUX especially runs
somewhat faster.

This also creates a new source file to hold the sound related code.