Replace JOYIN call with custom routine
Remove calls to the ROM joystick reading routine with a new routine. This
removes an indirect call in two places. The implementation currently used
is based on the ROM routine but has different behaviour in how it manages
the retries. Instead of trying ten times to get a value matching that which
was obtained on the previous call to GETJOY, it simply attempts to get two
matching values in a row. This may lead to jumpier joystick behaviour with
noisy pots so testing is needed. Of course, the routine can also be replaced
with any other useful routine if desired.