Add dependency tracking for building .bin files
Because main.asm includes other files in it, it is helpful to have
automatic tracking for dependencies so make can know it doesn't need
to rebuild some particular file when a source file not related to it
is modified. In the current tree, such a case would be the laser.asm
file being modified. The main build doesn't depend on that file so
there's no need to rebuild the main temple.bin file when laser.asm
is changed.